+++ packages/site/.gitignore
@@ -1,0 +1,3 @@
+# dumi2 website builder
+.dumi/tmp
+.dumi/tmp-production
