+++ packages/site/examples/creative/case/demo/meta.json
@@ -30,21 +30,6 @@
       "screenshot": "https://gw.alipayobjects.com/zos/rmsportal/neWvhxDYSxbNVSoZIgMN.gif"
     },
     {
-      "filename": "forecast.jsx",
-      "title": "预测线",
-      "screenshot": "https://gw.alipayobjects.com/mdn/rms_4f0ff1/afts/img/A*sDB-S6qxxsYAAAAAAAAAAABkARQnAQ"
-    },
-    {
-      "filename": "gauge.jsx",
-      "title": "半圆进度条",
-      "screenshot": "https://gw.alipayobjects.com/zos/rmsportal/fzCCwErtNDbjwJZCNFQR.gif"
-    },
-    {
-      "filename": "steps-record.jsx",
-      "title": "健康步数",
-      "screenshot": "https://gw.alipayobjects.com/zos/rmsportal/MLIlBSbtaBRiivHdBGmM.png"
-    },
-    {
       "filename": "radar.jsx",
       "title": "雷达图动画",
       "screenshot": "https://gw.alipayobjects.com/zos/finxbff/compress-tinypng/191d53e5-ba04-49f9-9119-f0a377723323.gif"
