+++ packages/site/examples/funnel/pyramid/demo/meta.json
@@ -5,7 +5,7 @@
   },
   "demos": [
     {
-      "filename": "basic.jsx",
+      "filename": "pyramid.jsx",
       "title": "基础金字塔图",
       "screenshot": "https://gw.alipayobjects.com/zos/finxbff/compress-tinypng/0ea9b22a-139b-4565-bc45-1dcec0b8e739.jpg"
     },
