--- packages/site/examples/gallery/index.en.md
@@ -1,7 +1,0 @@
----
-title: Gallery
-order: -1
-icon: other
-redirect_from:
-  - /en/examples
----
