--- packages/site/examples/gallery/index.zh.md
@@ -1,7 +1,0 @@
----
-title: 所有图表
-order: -1
-icon: other
-redirect_from:
-  - /zh/examples
----
