+++ packages/site/examples/pie/donut/demo/meta.json
@@ -20,11 +20,6 @@
       "screenshot": "https://gw.alipayobjects.com/mdn/rms_4f0ff1/afts/img/A*QlKgTb7bk3sAAAAAAAAAAABkARQnAQ"
     },
     {
-      "filename": "selected.jsx",
-      "title": "嵌套环图点击",
-      "screenshot": "https://gw.alipayobjects.com/zos/finxbff/compress-tinypng/aaee5282-92ca-49c9-9b36-25a5ab450567.jpg"
-    },
-    {
       "filename": "progress-bar.jsx",
       "title": "环形进度条",
       "screenshot": "https://gw.alipayobjects.com/mdn/rms_4f0ff1/afts/img/A*o56fS55cFGcAAAAAAAAAAABkARQnAQ"
