+++ packages/site/examples/pie/rose/demo/meta.json
@@ -5,7 +5,7 @@
   },
   "demos": [
     {
-      "filename": "rose.js",
+      "filename": "rose.jsx",
       "title": "玫瑰图",
       "screenshot": "https://gw.alipayobjects.com/mdn/rms_4f0ff1/afts/img/A*agBZRoedYKQAAAAAAAAAAABkARQnAQ"
     }
