+++ packages/site/package.json
@@ -11,24 +11,19 @@
   },
   "author": "https://github.com/orgs/antvis/people",
   "scripts": {
-    "site:build": "npm run site:clean && gatsby build --prefix-paths",
-    "site:clean": "gatsby clean",
+    "site:build": "dumi build",
+    "site:clean": "",
     "site:deploy": "npm run site:build && gh-pages -d public",
-    "site:develop": "gatsby develop --open",
+    "site:develop": "dumi dev",
     "site:start": "npm run site:develop"
-  },
-  "resolutions": {
-    "monaco-editor": "0.20.0",
-    "monaco-editor-webpack-plugin": "1.9.1",
-    "react-monaco-editor": "0.34.0"
   },
   "dependencies": {
     "@antv/f2": "^4.0.37",
     "@antv/f2-graphic": "^0.0.11",
     "@antv/f2-react": "^4.0.37",
-    "@antv/gatsby-theme-antv": "~1.1.15",
+    "@antv/dumi-theme-antv": "^0.3.0-beta.4",
+    "dumi": "^2.0.0-beta.15",
     "f2-touchemulator": "~0.0.1",
-    "gatsby": "^2.17.7",
     "gh-pages": "^2.1.1"
   },
   "devDependencies": {
