--- packages/site/site/locale.json
@@ -1,17 +1,0 @@
-{
-  "首页": "Home Page",
-  "蚂蚁数据可视化": "AntV",
-  "F2 移动端可视化方案": "F2, born for mobile",
-  "F2 是一个专注于移动，开箱即用的可视化解决方案，完美支持 H5 环境同时兼容多种环境（Node, 小程序，Weex），完备的图形语法理论，满足你的各种可视化需求，专业的移动设计指引为你带来最佳的移动端图表体验。": "F2 is born for mobile, developed for developers as well as designers. It is Html5 Canvas-based, and is also compatible with Node.js, Weex and React Native. Based on the grammar of graphics, F2 provides all the chart types you'll need. Our mobile design guidelines enable better user experience in mobile visualzation projects",
-  "开始使用": "Get Started",
-  "图表演示": "Examples",
-  "专注移动，体验优雅": "Focus on Mobile",
-  "围绕设计、性能以及异构环境，为用户提供移动端图表的最佳实践": "Best practices for moving side charts around design, performance and heterogeneous environments",
-  "图表丰富，组件完备": "Rich Chart",
-  "基于图形语法，可灵活构建各类图表（50+），组件完备，覆盖各类场景": "Based on the grammar of graphics, Flexible construction of various charts (50+), complete components, covering various scenarios",
-  "扩展灵活，创意无限": "Flexible & Infinite",
-  "插件机制，图形、动画、交互均可灵活扩展，使用更自由": "Plugin, graphics, animation and interaction can be flexibly expanded and freely used",
-  "蚂蚁财富": "Ant Fortune",
-  "支付宝": "Alipay",
-  "感谢信赖": "WE ARE TRUSTED BY"
-}
