--- packages/site/site/pages/index.en.tsx
@@ -1,3 +1,0 @@
-import Index from './index.zh';
-
-export default Index;
