--- packages/site/site/pages/index.less
@@ -1,3 +1,0 @@
-.teaser{
-  right: 0 !important;
-}
